refactor(editor): studio storage model 분리
